+++ modules/files/src/main/kotlin/org/archivekeep/files/repo/ObservableRepo.kt
@@ -1,10 +1,10 @@
 package org.archivekeep.files.repo
 
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.SharedFlow
+import kotlinx.coroutines.flow.StateFlow
 import org.archivekeep.utils.loading.Loadable
 
 interface ObservableRepo {
-    val indexFlow: SharedFlow<Loadable<RepoIndex>>
+    val indexFlow: StateFlow<Loadable<RepoIndex>>
     val metadataFlow: Flow<Loadable<RepositoryMetadata>>
 }