+++ modules/files/src/main/kotlin/org/archivekeep/files/repo/ObservableWorkingRepo.kt
@@ -2,7 +2,8 @@ package org.archivekeep.files.repo
 
 import kotlinx.coroutines.flow.Flow
 import org.archivekeep.files.operations.StatusOperation
+import org.archivekeep.utils.loading.Loadable
 
 interface ObservableWorkingRepo : ObservableRepo {
-    val localIndex: Flow<StatusOperation.Result>
+    val localIndex: Flow<Loadable<StatusOperation.Result>>
 }