+++ modules/utils/src/main/kotlin/org/archivekeep/utils/flows/logResourceLoad.kt
@@ -1,0 +1,6 @@
+package org.archivekeep.utils.flows
+
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.onEach
+
+fun <T> (Flow<T>).logResourceLoad(name: String): Flow<T> = this.onEach { println("Resource loaded: $name") }